+++ src/modules/account/currency-api.js
@@ -2,11 +2,12 @@
 
 import { Bridgeable } from 'yaob'
 
-import type {
-  EdgeCurrencyInfo,
-  EdgeCurrencyPlugin,
-  EdgeCurrencyTools
-} from '../../edge-core-index.js'
+import {
+  type EdgeCurrencyInfo,
+  type EdgeCurrencyPlugin,
+  type EdgeCurrencyTools,
+  type EdgeExchangeTools
+} from '../../index.js'
 import type { ApiInput } from '../root.js'
 import { changePluginSettings } from './account-files.js'
 
@@ -29,16 +30,50 @@ export class CurrencyTools extends Bridgeable<EdgeCurrencyTools> {
     return this._plugin.currencyInfo
   }
 
-  get pluginSettings (): Object {
+  get settings (): Object {
     return this._ai.props.state.currency.settings[this._plugin.pluginName]
   }
 
-  async changePluginSettings (settings: Object): Promise<mixed> {
+  async changeSettings (settings: Object): Promise<mixed> {
     await changePluginSettings(
       this._ai,
       this._accountId,
       this._plugin,
       settings
     )
+  }
+
+  // Deprecated names:
+  get pluginSettings (): Object {
+    return this.settings
+  }
+  async changePluginSettings (settings: Object): Promise<mixed> {
+    return this.changeSettings(settings)
+  }
+}
+
+export class ExchangeTools extends Bridgeable<EdgeExchangeTools> {
+  _settings: Object
+
+  constructor () {
+    super()
+    this._settings = {}
+  }
+
+  // TODO: Type EdgeExchangeInfo
+  get exchangeInfo (): Object {
+    return {
+      pluginName: 'shapeshift',
+      exchangeName: 'ShapeShift',
+      homepage: 'https://shapeshift.io/'
+    }
+  }
+
+  get settings (): Object {
+    return this._settings
+  }
+
+  async changeSettings (settings: Object): Promise<mixed> {
+    this._settings = settings
   }
 }