+++ src/modules/login/recovery2.js
@@ -3,10 +3,14 @@
 import { decrypt, encrypt, hmacSha256 } from '../../util/crypto/crypto.js'
 import { fixOtpKey, totp } from '../../util/crypto/hotp.js'
 import { base64, utf8 } from '../../util/encoding.js'
-import type { ApiInput } from '../root.js'
+import { type ApiInput } from '../root.js'
 import { authRequest } from './authServer.js'
 import { fixUsername, getStash } from './login-selectors.js'
-import type { LoginKit, LoginStash, LoginTree } from './login-types.js'
+import {
+  type LoginKit,
+  type LoginStash,
+  type LoginTree
+} from './login-types.js'
 import { applyKit, applyLoginReply, makeLoginTree } from './login.js'
 import { saveStash } from './loginStore.js'