+++ src/modules/scrypt/scrypt-pixie.js
@@ -1,10 +1,9 @@
 // @flow
 
-import type { PixieInput } from 'redux-pixies'
-import { combinePixies, stopUpdates } from 'redux-pixies'
+import { type PixieInput, combinePixies, stopUpdates } from 'redux-pixies'
 
 import { base16, utf8 } from '../../util/encoding.js'
-import type { RootProps } from '../root.js'
+import { type RootProps } from '../root.js'
 
 export type JsonSnrp = {
   salt_hex: string,
@@ -108,7 +107,7 @@ export function calcSnrpForTarget (
   return snrp
 }
 
-export default combinePixies({
+export const scrypt = combinePixies({
   makeSnrp: (input: PixieInput<RootProps>) => () => {
     const { io } = input.props
     let benchmark: Promise<number>