+++ test/modules/exchange/exchange.test.js
@@ -3,9 +3,11 @@
 import { assert, expect } from 'chai'
 import { describe, it } from 'mocha'
 
-import { makeFakeIos } from '../../../src/edge-core-index.js'
-import reducer from '../../../src/modules/exchange/exchange-reducer.js'
-import type { ExchangePair } from '../../../src/modules/exchange/exchange-reducer.js'
+import { makeFakeIos } from '../../../src/index.js'
+import {
+  type ExchangePair,
+  exchangeCache as reducer
+} from '../../../src/modules/exchange/exchange-reducer.js'
 import { getExchangeRate } from '../../../src/modules/exchange/exchange-selectors.js'
 import { makeCoreRoot } from '../../../src/modules/root.js'
 import {